continuing size_t infiltration
svn path=/trunk/yarp2/; revision=8918

+++ src/carriers/human_carrier/HumanCarrier.h
@@ -61,7 +61,7 @@ public:
 
     virtual void getHeader(const Bytes& header) {
         const char *target = "HUMANITY";
-        for (int i=0; i<8 && i<header.length(); i++) {
+        for (size_t i=0; i<8 && i<header.length(); i++) {
             header.get()[i] = target[i];
         }
     }
@@ -71,7 +71,7 @@ public:
             return false;
         }
         const char *target = "HUMANITY";
-        for (int i=0; i<8; i++) {
+        for (size_t i=0; i<8; i++) {
             if (header.get()[i] != target[i]) {
                 return false;
             }
