+++ src/carriers/mpi_carrier/src/MpiCarrier.cpp
@@ -29,7 +29,7 @@ MpiCarrier::~MpiCarrier() {
 }
 
 void  MpiCarrier::getHeader(const Bytes& header) {
-    for (int i=0; i<8 && i<header.length(); i++) {
+    for (size_t i=0; i<8 && i<header.length(); i++) {
         header.get()[i] = target.c_str()[i];
     }
 }
