+++ src/carriers/tcpros_carrier/TcpRosCarrier.h
@@ -114,7 +114,7 @@ public:
     virtual void getHeader(const Bytes& header) {
         // no header, will need to do some fancy footwork
         const char *target = "NONONONO";
-        for (int i=0; i<8 && i<header.length(); i++) {
+        for (size_t i=0; i<8 && i<header.length(); i++) {
             header.get()[i] = target[i];
         }
     }
